fix(accordion): move item props to context
Before this change, `Accordion` passed the `isOpen` and `onChange` props
to its `AccordionItem` children by injecting its props via
`React.cloneElement`. This method of props injection caused two issues:

- `AccordionItem` children could only be rendered as direct children of
  `Accordion`
- `AccordionItem` children rendered as a mapped array could not receive
  their corresponding props, breaking the functionality of the accordion

This changes moves the `isOpen` and `onChange` props into a props getter
created in `useAccordion`. This props getter takes an accordion item's
`index` (determined in `useAccordionItem` via `useDescendant`) to
determine these values. This props getter is then attached to
`AccordionContext`.

The props getter is then accessed via `AccordionContext` inside
`useAccordionItem` and passed the item's index. This allows the item to
function anywhere inside the `Accordion` tree, whether as direct
children or not, and as part of a mapped array.